test: specific test for failed multichain call
One of the MythX reports highlighted that failures in multicall in were
not tracked or reverted. The lack of revert was on purpose and is a
break away from the way ENS handles things. This removes safety if
multiple "write" functions are called because some writes will make it
through but others could fail. The intent was to improve "read"
functions by being able to send many read functions and let them fail,
handling missing data on the caller side.